finished API

+++ models/meetgreet.js
@@ -29,11 +29,11 @@ module.exports = (sequelize, DataTypes) => {
       autoIncrement: true
     },
     event_id: {
-      type: DataTypes.SMALLINT,
+      type: DataTypes.INTEGER,
       allowNull: false
     },
     band_id: {
-      type: DataTypes.SMALLINT,
+      type: DataTypes.INTEGER,
       allowNull: false
     },
     meet_start_time: {
